Decompose controller unit tests
